feat: add system roles on project creation

+++ apps/server/src/resources/project-role/business.ts
@@ -6,7 +6,7 @@ import {
   listRoles as listRolesQuery,
   updateRole,
 } from '@/resources/queries-index.js'
-import { BadRequest400, NotFound404 } from '@/utils/errors.js'
+import { BadRequest400, Forbidden403, NotFound404 } from '@/utils/errors.js'
 import prisma from '@/prisma.js'
 
 export async function listRoles(projectId: Project['id']) {
@@ -31,6 +31,9 @@ export async function patchRoles(projectId: Project['id'], roles: typeof project
       if (typeof matchingRole?.position !== 'undefined' && !positionsAvailable.includes(matchingRole.position)) {
         positionsAvailable.push(matchingRole.position)
       }
+      if (dbRole.type === 'system') {
+        throw new Forbidden403('Ce rôle système ne peut pas être renommé')
+      }
       if (matchingRole?.oidcGroup && !matchingRole.oidcGroup.startsWith('/')) {
         throw new BadRequest400('oidcGroup doit commencer par /')
       }
@@ -40,6 +43,7 @@ export async function patchRoles(projectId: Project['id'], roles: typeof project
         permissions: matchingRole?.permissions ? BigInt(matchingRole?.permissions) : BigInt(dbRole.permissions),
         position: matchingRole?.position ?? dbRole.position,
         oidcGroup: matchingRole?.oidcGroup ? `/${project.slug}${matchingRole.oidcGroup}` : dbRole.oidcGroup,
+        type: matchingRole?.type ?? dbRole.type,
       }
     })
   if (positionsAvailable.length && positionsAvailable.length !== dbRoles.length) return new BadRequest400('Les numéros de position des rôles sont incohérentes')
@@ -58,6 +62,10 @@ export async function createRole(projectId: Project['id'], role: typeof projectR
     orderBy: { position: 'desc' },
     select: { position: true },
   }))?.position ?? -1
+
+  if (role.type === 'system') {
+    throw new Forbidden403('Ce rôle système ne peut pas être renommé')
+  }
 
   if (role.oidcGroup && !role.oidcGroup.startsWith('/')) {
     throw new BadRequest400('oidcGroup doit commencer par /')
@@ -89,6 +97,10 @@ export async function countRolesMembers(projectId: Project['id']) {
 }
 
 export async function deleteRole(roleId: Project['id']) {
+  const role = await prisma.projectRole.findUnique({ where: { id: roleId } })
+  if (role?.type === 'system') {
+    throw new Forbidden403('Ce rôle système ne peut pas être supprimé')
+  }
   await deleteRoleQuery(roleId)
   return null
 }
